Bug fixes, added logging to dump, move set_proxy_user_password to globals.py. Renamed common.py to globals.py
