+++ src/commodity_data/commodity_data.py
@@ -1,8 +1,8 @@
 import pandas as pd
 
-from commodity_data.common import logger
 from commodity_data.downloaders import EEXDownloader, OmipDownloader, BarchartDownloader
 from commodity_data.downloaders.base_downloader import BaseDownloader, FilterKeyNotFoundException
+from commodity_data.globals import logger
 
 
 class CommodityData:
