+++ src/commodity_data/downloaders/barchart/barchart_data.py
@@ -9,9 +9,9 @@ import numpy as np
 import pandas as pd
 from ong_utils import get_cookies
 
-from commodity_data.common import logger
 from commodity_data.downloaders.base_downloader import HttpGet
 from commodity_data.downloaders.series_config import TypeColumn
+from commodity_data.globals import logger
 
 
 class BarchartData(HttpGet):
