+++ src/commodity_data/downloaders/omip/omip_data.py
@@ -2,9 +2,9 @@ from datetime import datetime
 
 import pandas as pd
 
-from commodity_data.common import logger
 from commodity_data.downloaders.base_downloader import HttpGet
 from commodity_data.downloaders.products import date_offset
+from commodity_data.globals import logger
 
 
 def parse_omip_product_maturity_offset(omip_product: str, as_of: pd.Timestamp) -> tuple:
